+++ CHANGELOG.md
@@ -13,7 +13,8 @@
   
   Dependency Upgrade
     * Upgrade Sundrio to 0.16.5
-  
+    * Upgrade to Bean Validation 2.0
+
   New Feature
 
 #### 4.1.3