+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   Bugs
     * Fix #1387: ValidatingWebhookConfigurationOperationsImpl should be a NonNamespaceOperation
     * Fix #1429: Fixes JsonMappingException: No resource type found for:v1#List when reading a Kubernetes List YAML
+    * Fix #760: Api get pod from yaml issue
+    * Fix #807: Loading a deployment from server with a config file throws exception
    
   Improvements