+++ CHANGELOG.md
@@ -14,7 +14,7 @@
     * Fix #1425: Preserve labels and fields when using CRD's withResourceVersion()
   
   Dependency Upgrade
-    * Upgrade Sundrio to 0.16.5
+    * Upgrade Sundrio to 0.17.0
     * Upgrade to Bean Validation 2.0
 
   New Feature